pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-trading"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "anthropic>=0.49",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
engine = "engine.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["engine*"]
